Refactor xtask util helpers into modular modules
